Machine learning/requirements.txt
langchain-core>=0.1.0
langchain-ollama>=0.1.0
flask>=2.0.0
python-dotenv>=1.0.0